UPDATE_SNAPSHOTS=true make test
