+++ pkg/grapicmd/internal/module/generator/.snapshots/Test_ServiceGenerator-book#01-Generate_without_test-app-server-book_server.go
@@ -3,9 +3,7 @@ package server
 import (
 	"context"
 
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/izumin5210/grapi/pkg/grapiserver"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -21,15 +19,5 @@ func NewBookServiceServer() interface {
 }
 
 type bookServiceServerImpl struct {
-}
-
-// RegisterWithServer implements grapiserver.Server.RegisterWithServer.
-func (s *bookServiceServerImpl) RegisterWithServer(grpcSvr *grpc.Server) {
-	api_pb.RegisterBookServiceServer(grpcSvr, s)
-}
-
-// RegisterWithHandler implements grapiserver.Server.RegisterWithHandler.
-func (s *bookServiceServerImpl) RegisterWithHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
-	return api_pb.RegisterBookServiceHandler(ctx, mux, conn)
 }
