+++ pkg/grapicmd/internal/module/generator/.snapshots/Test_ServiceGenerator-book#01-Generate_without_test-app-server-book_server_register_funcs.go
@@ -1,1 +1,21 @@
+package server
+
+import (
+	"context"
+
+	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+	"google.golang.org/grpc"
+
+	api_pb "testapp/api"
+)
+
+// RegisterWithServer implements grapiserver.Server.RegisterWithServer.
+func (s *bookServiceServerImpl) RegisterWithServer(grpcSvr *grpc.Server) {
+	api_pb.RegisterBookServiceServer(grpcSvr, s)
+}
+
+// RegisterWithHandler implements grapiserver.Server.RegisterWithHandler.
+func (s *bookServiceServerImpl) RegisterWithHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
+	return api_pb.RegisterBookServiceHandler(ctx, mux, conn)
+}
