+++ pkg/grapicmd/internal/module/generator/.snapshots/Test_ServiceGenerator-foo-bar-baz-Generate-app-server-foo-bar_baz_server.go
@@ -3,9 +3,7 @@ package foo
 import (
 	"context"
 
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/izumin5210/grapi/pkg/grapiserver"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -21,15 +19,5 @@ func NewBarBazServiceServer() interface {
 }
 
 type barBazServiceServerImpl struct {
-}
-
-// RegisterWithServer implements grapiserver.Server.RegisterWithServer.
-func (s *barBazServiceServerImpl) RegisterWithServer(grpcSvr *grpc.Server) {
-	foo_pb.RegisterBarBazServiceServer(grpcSvr, s)
-}
-
-// RegisterWithHandler implements grapiserver.Server.RegisterWithHandler.
-func (s *barBazServiceServerImpl) RegisterWithHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
-	return foo_pb.RegisterBarBazServiceHandler(ctx, mux, conn)
 }
