+++ pkg/grapicmd/internal/module/generator/.snapshots/Test_ServiceGenerator-foo-bar-baz_with_list,create,delete-Generate-app-server-foo-bar_baz_server_register_funcs.go
@@ -1,1 +1,21 @@
+package foo
+
+import (
+	"context"
+
+	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+	"google.golang.org/grpc"
+
+	foo_pb "testapp/api/foo"
+)
+
+// RegisterWithServer implements grapiserver.Server.RegisterWithServer.
+func (s *barBazServiceServerImpl) RegisterWithServer(grpcSvr *grpc.Server) {
+	foo_pb.RegisterBarBazServiceServer(grpcSvr, s)
+}
+
+// RegisterWithHandler implements grapiserver.Server.RegisterWithHandler.
+func (s *barBazServiceServerImpl) RegisterWithHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
+	return foo_pb.RegisterBarBazServiceHandler(ctx, mux, conn)
+}
