+++ pkg/grapicmd/internal/module/generator/.snapshots/Test_ServiceGenerator-foo-bar-baz_with_list,create,rename,delete,move_move-Generate-app-server-foo-bar_baz_server.go
@@ -4,9 +4,7 @@ import (
 	"context"
 
 	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/izumin5210/grapi/pkg/grapiserver"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -22,16 +20,6 @@ func NewBarBazServiceServer() interface {
 }
 
 type barBazServiceServerImpl struct {
-}
-
-// RegisterWithServer implements grapiserver.Server.RegisterWithServer.
-func (s *barBazServiceServerImpl) RegisterWithServer(grpcSvr *grpc.Server) {
-	foo_pb.RegisterBarBazServiceServer(grpcSvr, s)
-}
-
-// RegisterWithHandler implements grapiserver.Server.RegisterWithHandler.
-func (s *barBazServiceServerImpl) RegisterWithHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
-	return foo_pb.RegisterBarBazServiceHandler(ctx, mux, conn)
 }
 
 func (s *barBazServiceServerImpl) ListBarBazs(ctx context.Context, req *foo_pb.ListBarBazsRequest) (*foo_pb.ListBarBazsResponse, error) {
